+++ src/events/mod.rs
@@ -673,37 +673,25 @@ pub struct BytesText<'a> {
 }
 
 impl<'a> BytesText<'a> {
-    /// Creates a new `BytesText` from an escaped byte sequence.
-    #[inline]
-    pub fn from_escaped<C: Into<Cow<'a, [u8]>>>(content: C) -> Self {
-        Self {
-            content: content.into(),
-        }
-    }
-
-    /// Creates a new `BytesText` from a byte sequence. The byte sequence is
-    /// expected not to be escaped.
-    #[inline]
-    pub fn from_plain(content: &'a [u8]) -> Self {
-        Self {
-            content: escape(content),
-        }
-    }
-
     /// Creates a new `BytesText` from an escaped string.
     #[inline]
-    pub fn from_escaped_str<C: Into<Cow<'a, str>>>(content: C) -> Self {
-        Self::from_escaped(match content.into() {
-            Cow::Owned(o) => Cow::Owned(o.into_bytes()),
-            Cow::Borrowed(b) => Cow::Borrowed(b.as_bytes()),
-        })
+    pub fn from_escaped<C: Into<Cow<'a, str>>>(content: C) -> Self {
+        let content = match content.into() {
+            Cow::Borrowed(c) => Cow::Borrowed(c.as_bytes().into()),
+            Cow::Owned(c) => Cow::Owned(c.into()),
+        };
+        Self { content }
     }
 
     /// Creates a new `BytesText` from a string. The string is expected not to
     /// be escaped.
     #[inline]
-    pub fn from_plain_str(content: &'a str) -> Self {
-        Self::from_plain(content.as_bytes())
+    pub fn from_plain(content: &'a str) -> Self {
+        let content = match escape(content).into() {
+            Cow::Borrowed(c) => Cow::Borrowed(c.as_bytes().into()),
+            Cow::Owned(c) => Cow::Owned(c.into()),
+        };
+        Self { content }
     }
 
     /// Ensures that all data is owned to extend the object's lifetime if
@@ -756,8 +744,7 @@ impl<'a> BytesText<'a> {
         &self,
         resolve_entity: impl Fn(&str) -> Option<&'entity str>,
     ) -> Result<Cow<str>> {
-        // from_utf8 should never fail because content is always UTF-8 encoded
-        Ok(unescape_with(from_utf8(&self.content)?, resolve_entity)?)
+        Ok(unescape_with(from_utf8(self)?, resolve_entity)?) // TODO(dalley)
     }
 
     /// Decodes then unescapes the content of the event.
@@ -902,10 +889,8 @@ impl<'a> BytesCData<'a> {
     /// | `'`       | `&apos;`
     /// | `"`       | `&quot;`
     pub fn escape(self) -> BytesText<'a> {
-        BytesText::from_escaped(match escape(&self.content) {
-            Cow::Borrowed(_) => self.content,
-            Cow::Owned(escaped) => Cow::Owned(escaped),
-        })
+        let content = std::str::from_utf8(&self.content).unwrap(); // TODO(dalley): this is temporary
+        BytesText::from_escaped(escape(&content)).into_owned()
     }
 
     /// Converts this CDATA content to an escaped version, that can be written
@@ -922,10 +907,8 @@ impl<'a> BytesCData<'a> {
     /// | `>`       | `&gt;`
     /// | `&`       | `&amp;`
     pub fn partial_escape(self) -> BytesText<'a> {
-        BytesText::from_escaped(match partial_escape(&self.content) {
-            Cow::Borrowed(_) => self.content,
-            Cow::Owned(escaped) => Cow::Owned(escaped),
-        })
+        let content = std::str::from_utf8(&self.content).unwrap(); // TODO(dalley): this is temporary
+        BytesText::from_escaped(partial_escape(&content)).into_owned()
     }
 
     /// Gets content of this text buffer in the specified encoding