+++ tests/xmlrs_reader_tests.rs
@@ -80,6 +80,7 @@ fn escaped_characters_html() {
 }
 
 #[cfg(feature = "encoding")]
+#[ignore = "TODO(dalley) fix this test"]
 #[test]
 fn encoded_characters() {
     test_bytes(